+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
  * [#173](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/173) - `autotag.legacyTags()` NPE when no arguments provided
+ * [#175](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/pull/175) - Remove empty commands before doing COPY consolidation
 
 ## Version 5.1.0 - 2020-10-20